+++ cpp/include/milvus-storage/filesystem/s3/multi_part_upload_s3_fs.h
@@ -24,8 +24,9 @@
 #include <aws/core/http/HttpClientFactory.h>
 
 #include <arrow/util/key_value_metadata.h>
-#include "arrow/filesystem/filesystem.h"
-#include "arrow/io/interfaces.h"
+#include <arrow/filesystem/filesystem.h>
+#include <arrow/io/interfaces.h>
+
 #include "milvus-storage/common/constants.h"
 #include "milvus-storage/filesystem/s3/s3_options.h"
 #include "milvus-storage/filesystem/s3/s3_client.h"
@@ -104,47 +105,5 @@ class MultiPartUploadS3FS : public arrow::fs::FileSystem {
   class Impl;
   std::shared_ptr<Impl> impl_;
 };
-
-// TODO: should move this logical out
-/// \brief Initialize the S3 APIs with the specified set of options.
-///
-/// It is required to call this function at least once before using S3FileSystem.
-///
-/// Once this function is called you MUST call FinalizeS3 before the end of the
-/// application in order to avoid a segmentation fault at shutdown.
-arrow::Status InitializeS3(const S3GlobalOptions& options);
-
-/// \brief Ensure the S3 APIs are initialized, but only if not already done.
-///
-/// If necessary, this will call InitializeS3() with some default options.
-arrow::Status EnsureS3Initialized();
-
-/// Whether S3 was initialized, and not finalized.
-bool IsS3Initialized();
-
-/// Whether S3 was finalized.
-bool IsS3Finalized();
-
-/// \brief Check if S3 is initialized and return an error if not.
-///
-/// This function checks if S3 has been initialized and returns an appropriate
-/// error status if it has not been initialized or has been finalized.
-arrow::Status CheckS3Initialized();
-
-/// \brief Shutdown the S3 APIs.
-///
-/// This can wait for some S3 concurrent calls to finish so as to avoid
-/// race conditions.
-/// After this function has been called, all S3 calls will fail with an error.
-///
-/// Calls to InitializeS3() and FinalizeS3() should be serialized by the
-/// application (this also applies to EnsureS3Initialized() and
-/// EnsureS3Finalized()).
-arrow::Status FinalizeS3();
-
-/// \brief Ensure the S3 APIs are shutdown, but only if not already done.
-///
-/// If necessary, this will call FinalizeS3().
-arrow::Status EnsureS3Finalized();
 
 }  // namespace milvus_storage
