+++ cpp/test/packed/packed_test_base.h
@@ -13,7 +13,6 @@
 // limitations under the License.
 
 #include <arrow/filesystem/filesystem.h>
-#include <arrow/filesystem/s3fs.h>
 #include <arrow/filesystem/localfs.h>
 #include <arrow/api.h>
 #include <arrow/array/builder_binary.h>
